requirements.txt
asgiref==3.8.1
boto3==1.35.59
botocore==1.35.59
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
django-selectel-storage==1.0.2
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.10
jmespath==1.0.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
